requirements.txt
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.3
scikit_learn==1.1.2
scipy==1.9.0
tabulate==0.8.10
scikit-fuzzy==0.5.0
pyyaml==6.0.2
networkx==2.8.6